custom_views/requirements.txt
django==1.10.5
django-material==0.12.5
viewflow=0.12.2

